refactor: move 29 remaining tests from maze to maze-javafx-backend with correct package namespaces
- Moved all 29 test files from maze/src/test/java/main/game/maze/ to their
  architecturally correct locations in maze-javafx-backend/src/test/java/
- Package declarations updated to match class-under-test locations:
    main.game.maze            -> maze-javafx-backend (unchanged, 13 files)
    main.game.maze            -> main.game.maze.characters  (7 files)
    main.game.maze.characters -> main.game.maze.characters  (CollisionDamageTest, unchanged)
    main.game.maze            -> main.game.maze.service     (1 file)
    main.game.maze            -> main.game.maze.actions     (3 files)
    main.game.maze            -> main.game.maze.areas       (1 file)
    main.game.maze            -> main.game.maze.behaviour   (2 files: PatrolBehaviorTest, TestPatrolBehaviorMoveManual)
- Added missing imports (main.game.maze.App, main.game.maze.GameController)
  to DeathSubscriptionUsageTest and GhostTangibilityTest after package change
- Added RestartGameActionTest import for main.game.maze.App
- maze/src/test/ is now empty (pure launcher module)
- Added scores.txt as a tracked test resource in maze-javafx-backend/src/test/resources/
  (force-added past gitignore) for HighScoreControllerTest
- Configured surefire workingDirectory to target/test-classes so test
  resources (scores.txt) are found by file-relative reads in tests
- All 113 tests pass (2 pre-existing @Disabled skips)
