+++ maze-javafx-backend/src/test/java/main/game/maze/GameControllerPathHintBudgetTest.java
