+++ maze-javafx-backend/src/test/java/main/game/maze/MazeCoreBehaviorTest.java
