+++ maze-javafx-backend/src/test/java/main/game/maze/behaviour/TestPatrolBehaviorMoveManual.java
@@ -1,4 +1,4 @@
-package main.game.maze;
+package main.game.maze.behaviour;
 
 import main.game.maze.behaviour.BehaviourFactory;
 import main.game.maze.behaviour.DijkstraPathCalculator;
