+++ maze-javafx-backend/src/test/java/main/game/maze/characters/CharacterDisposeTest.java
@@ -1,4 +1,4 @@
-package main.game.maze;
+package main.game.maze.characters;
 
 import javafx.application.Platform;
 import javafx.scene.Node;
