+++ maze-javafx-backend/src/test/java/main/game/maze/characters/CollisionDamageTest.java
