+++ maze-javafx-backend/src/test/resources/scores.txt
@@ -1,0 +1,3 @@
+KingKong: 5090
+YesMan: 2990
+Neddis: 1520
